Rewrote part of +hurt
Coverage tests revealed the ability to get into an loop because of how damage upgrading was working. Also tweaked the messaging a bit since it ends up being possible to take multiple forms of damage at the same time thanks to upgrading (e.g. character takes a large amount of bashing and a few points upgrade to lethal)